Fix executing commands on Android chromium browsers
In Chromium-based browsers on Android, when there are multiple input fields and the `enterkeyhint` attribute is not set, pressing the Enter key on the virtual keyboard will automatically switch to the next input field.

This behavior contradicts our expectation of inserting a line break.

Therefore, we need to manually set the enterkeyhint attribute to 'enter' in the input fields.

https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/enterkeyhint